backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.0

# Pydantic
pydantic==2.10.0
pydantic-settings==2.7.0

# Auth
python-jose[cryptography]==3.3.0
cryptography==44.0.0
httpx==0.27.2

# Task Queue
celery==5.4.0
redis==5.2.0

# AI - LiteLLM gateway
litellm==1.55.0

# AI - LangGraph agents
langgraph==0.2.60

# AI - Observability
langfuse==2.57.0

# Object Storage (S3-compatible for R2)
boto3==1.35.0

# Utilities
orjson==3.10.12
python-dateutil==2.9.0
